added comments and fixed pep8
Added CLI options to support comments, by giving a query, the comment
and if it should CC the users manager. The CC is based on the assumption
that LDAP is being used, and the LDAP server is also given as a parm.

Pep8 violations were also fixed in this commit